Add files via upload
I am proud of developing the KNN, Naive Bayes, and Decision Tree models for the Fetal Health Classification Project. Reduction of child mortality is reflected in several of the United Nations Sustainable Development Goals and is a key indicator of human progress. The UN expects that by 2030, countries aim to reduce under-5 mortality to at least as low as 25 per 1000 live births, emphasizing the need for proactive measures.

Cardiotocograms (CTGs) serve as a simple and cost-effective means to assess fetal health, enabling healthcare professionals to intervene effectively and prevent child and maternal mortality. The project's challenges included selecting appropriate machine learning algorithms, handling large datasets, and ensuring model accuracy in classifying fetal health as Normal, Suspect, or Pathological. It's a project that aligns with critical global health goals and underscores the potential of AI in healthcare.